Issue #18535: Move list/grid viewholders to browser package
