+++ app/src/main/java/org/mozilla/fenix/tabstray/browser/BrowserTabsAdapter.kt
@@ -16,8 +16,6 @@ import mozilla.components.support.base.observer.Observable
 import mozilla.components.support.base.observer.ObserverRegistry
 import org.mozilla.fenix.ext.components
 import org.mozilla.fenix.ext.settings
-import org.mozilla.fenix.tabstray.TabsTrayGridViewHolder
-import org.mozilla.fenix.tabstray.TabsTrayListViewHolder
 import org.mozilla.fenix.tabstray.TabsTrayViewHolder
 
 /**
